Move thrift-to-tag into malloy
